stockwatch/requirements.txt
celery==5.5.0
channels==4.2.2
Django==4.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pymongo==3.11.4
python-decouple==3.8
Requests==2.32.3
djongo==1.3.7
sqlparse==0.2.4